+++ src/lib/public-api.ts
@@ -17,8 +17,6 @@ export { TerraButtonComponent } from './components/buttons/button/terra-button.c
 export { TerraButtonWithOptionsComponent } from './components/buttons/button-with-options/terra-button-with-options.component';
 export { TerraToggleComponent } from './components/buttons/toggle/terra-toggle.component';
 
-export { TerraFileChooserComponent } from './components/buttons/file-chooser/terra-file-chooser.component';
-
 export { TerraBaseData } from './components/data/terra-base.data';
 export { FilterComponent } from './components/filter/filter.component';