+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Breaking Changes
 
 -   **terra-info** has been removed.
+-   **terra-file-chooser** has been removed.
+-   **terra-file-input** replaced `terra-file-chooser` with the button & dialog component provided by `@angular/material`.
 
 # 12.2.0 (X.X.2021)